+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## Unreleased
-- Add BaseCharacter class and tests (not yet implemented) [#17](https://github.com/jeromecovington/sysref-base/pull/17)
-- Refactor bestiary provider to BaseMonster class [#16](https://github.com/jeromecovington/sysref-base/pull/16)
+- Export `BaseMonster` class [#19](https://github.com/jeromecovington/sysref-base/pull/19)
+- Add `BaseCharacter` class and tests (not yet implemented) [#17](https://github.com/jeromecovington/sysref-base/pull/17)
+- Refactor bestiary provider to `BaseMonster` class [#16](https://github.com/jeromecovington/sysref-base/pull/16)
 - Implement xp in bestiary [#15](https://github.com/jeromecovington/sysref-base/pull/15)
 
 ## 0.4.0